build(psm-tool): clasp: google apps script dev tool
chore(psm-tool): sync with Google sheet

chore(psm-tool): move psm-tool.js to src/

chore(psm-tool): split between script and lib

chore(psm-tool): rename Code.js from google sheets

refactor(psm-tool): move CLI details out of src/psm-lib.js

chore(psm-tool): move Google apps scripts stuff to app/

fixup(psm-tool): duplicate assert in sheets-lib.js

feat(psm-tool): simpleWalletState for use in Google Sheets

 - theWeb

fix(psm-tool): avoid xs.last() in Google Apps Script

also:
 - 1_000_000
 - 123n

build(psm-tool): build, clasp:push

chore(psm-tool): ignore google apps scripts config
