+++ packages/psm-tool/.gitignore
@@ -1,0 +1,1 @@
+app/.clasp.json
