+++ packages/psm-tool/app/appsscript.json
@@ -1,0 +1,7 @@
+{
+  "timeZone": "America/Chicago",
+  "dependencies": {
+  },
+  "exceptionLogging": "STACKDRIVER",
+  "runtimeVersion": "V8"
+}
