+++ packages/psm-tool/package.json
@@ -27,7 +27,8 @@
     "./package.json": "./package.json"
   },
   "scripts": {
-    "build": "exit 0",
+    "build": "mkdir -p app/build; sed 's/^export .*//' src/psm-lib.js >app/build/psm-lib.js",
+    "clasp:push": "yarn build && cd app; clasp push",
     "lint": "yarn lint:types && yarn lint:js",
     "lint-fix": "eslint --fix .",
     "lint:js": "eslint .",
@@ -37,6 +38,7 @@
   "dependencies": {},
   "devDependencies": {
     "@endo/eslint-config": "^0.3.6",
+    "@google/clasp": "^2.4.1",
     "ava": "^3.12.1",
     "babel-eslint": "^10.0.3",
     "eslint": "^7.23.0",
